Crate panic_semihosting

Report panic messages to the host stderr using semihosting

This crate contains an implementation of panic_fmt that logs panic messages to the host stderr using cortex-m-semihosting. Before logging the message the panic handler disables (masks) the device specific interrupts. After logging the message the panic handler trigger a breakpoint and then goes into an infinite loop.

Currently, this crate only supports the ARM Cortex-M architecture.

Usage

#![no_std]

extern crate panic_semihosting;

fn main() {
    panic!("FOO")
}

(gdb) monitor arm semihosting enable
(gdb) continue
Program received signal SIGTRAP, Trace/breakpoint trap.
rust_begin_unwind (args=..., file=..., line=8, col=5)
    at $CRATE/src/lib.rs:69
69          asm::bkpt();

$ openocd -f (..)
(..)
panicked at 'FOO', src/main.rs:6:5

Optional features

inline-asm

When this feature is enabled semihosting is implemented using inline assembly (asm!) and compiling this crate requires nightly.

When this feature is disabled semihosting is implemented using FFI calls into an external assembly file and compiling this crate works on stable and beta.